# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/translate/config": {
      "get": {
        "operationId": "TranslateController_getConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current translation config"
          }
        },
        "summary": "Get runtime translation config",
        "tags": [
          "translate"
        ]
      },
      "put": {
        "operationId": "TranslateController_updateConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTranslateConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated translation config"
          }
        },
        "summary": "Update runtime translation config (applied immediately, no restart)",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/llm/test": {
      "post": {
        "operationId": "TranslateController_testLlm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmProbeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "{ ok, message }"
          }
        },
        "summary": "Test-connect to an LLM endpoint with the given (possibly unsaved) params",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/llm/models": {
      "post": {
        "operationId": "TranslateController_listLlmModels",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmProbeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "{ models: string[] }"
          }
        },
        "summary": "List available model names for an LLM endpoint (Ollama/OpenAI-compatible)",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/preview": {
      "post": {
        "operationId": "TranslateController_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewTranslateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "{ pair, translated }"
          }
        },
        "summary": "Translate ad-hoc text through the live pipeline (glossary/senders/casing)",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/glossary": {
      "get": {
        "operationId": "TranslateController_getGlossary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Glossary terms (source = 中文, target = 越南文)"
          }
        },
        "summary": "List zh<->vi glossary terms",
        "tags": [
          "translate"
        ]
      },
      "post": {
        "operationId": "TranslateController_addGlossary",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlossaryTermDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Updated glossary terms"
          }
        },
        "summary": "Add/overwrite a zh<->vi glossary term (both directions)",
        "tags": [
          "translate"
        ]
      },
      "delete": {
        "operationId": "TranslateController_removeGlossary",
        "parameters": [
          {
            "name": "term",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated glossary terms"
          }
        },
        "summary": "Remove any glossary pairing where the term appears on either side",
        "tags": [
          "translate"
        ]
      }
    },
...
    "/api/translate/glossary/pending/{id}/approve": {
      "post": {
        "operationId": "TranslateController_approvePendingGlossary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Remaining pending suggestions"
          }
        },
        "summary": "Approve a pending glossary suggestion into the glossary",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/glossary/pending/{id}": {
      "delete": {
        "operationId": "TranslateController_rejectPendingGlossary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remaining pending suggestions"
          }
        },
        "summary": "Reject (drop) a pending glossary suggestion",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/memory/candidates": {
      "get": {
        "operationId": "TranslateController_getMemoryCandidates",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidates ordered by frequency"
          }
        },
        "summary": "Top translation-memory candidates to promote into the glossary",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/memory/{id}/approve": {
      "post": {
        "operationId": "TranslateController_approveMemoryCandidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Remaining candidates"
          }
        },
        "summary": "Promote a memory candidate into the glossary",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/memory/{id}": {
      "delete": {
        "operationId": "TranslateController_dismissMemoryCandidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remaining candidates"
          }
        },
        "summary": "Dismiss a memory candidate",
        "tags": [
          "translate"
        ]
      }
    },
...
    "/api/translate/memory/phrases": {
      "get": {
        "operationId": "TranslateController_getPhraseCandidates",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Phrase candidates ordered by frequency"
          }
        },
        "summary": "High-frequency phrase candidates awaiting review",
        "tags": [
          "translate"
        ]
      }
    },
...
    "/api/translate/glossary/overrides": {
      "get": {
        "operationId": "TranslateController_getGlossaryOverrides",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Overrides grouped by scope"
          }
        },
        "summary": "Per-group glossary overrides, their counts, and orphaned layers",
        "tags": [
          "translate"
        ]
      },
      "delete": {
        "operationId": "TranslateController_removeGlossaryOverride",
        "parameters": [
          {
            "name": "group",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "term",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Refreshed overrides"
          }
        },
        "summary": "Drop one group's override so the shared glossary term applies again",
        "tags": [
          "translate"
        ]
      }
    },
...
    "/api/translate/memory/phrases/approve-bulk": {
      "post": {
        "operationId": "TranslateController_approvePhrasesBulk",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveBulkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "How many moved, out of how many matched"
          }
        },
        "summary": "Approve every candidate at or above a frequency threshold (max 200)",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/memory/phrases/{id}/approve": {
      "post": {
        "operationId": "TranslateController_approvePhraseCandidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Remaining phrase candidates"
          }
        },
        "summary": "Promote a phrase candidate into the glossary",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/memory/phrases/{id}": {
      "delete": {
        "operationId": "TranslateController_dismissPhraseCandidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remaining phrase candidates"
          }
        },
        "summary": "Dismiss a phrase candidate",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/senders": {
      "get": {
        "operationId": "TranslateController_getSenders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sender overrides"
          }
        },
        "summary": "List @mention JID->name overrides",
        "tags": [
          "translate"
        ]
      },
      "post": {
        "operationId": "TranslateController_addSender",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SenderEntryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Updated sender overrides"
          }
        },
        "summary": "Add/overwrite an @mention JID->name override",
        "tags": [
          "translate"
        ]
      },
      "delete": {
        "operationId": "TranslateController_removeSender",
        "parameters": [
          {
            "name": "jid",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated sender overrides"
          }
        },
        "summary": "Remove an @mention JID override",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/categories": {
      "get": {
        "operationId": "TranslateController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Categories"
          }
        },
        "summary": "List glossary categories",
        "tags": [
          "translate"
        ]
      },
      "post": {
        "operationId": "TranslateController_addCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Updated categories"
          }
        },
        "summary": "Add a glossary category",
        "tags": [
          "translate"
        ]
      },
      "delete": {
        "operationId": "TranslateController_removeCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated categories"
          }
        },
        "summary": "Remove a glossary category",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/senders/backfill": {
      "post": {
        "operationId": "TranslateController_backfillSenders",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSendersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Count filled + updated overrides"
          }
        },
        "summary": "Resolve names for queued empty-name @lid entries via the session contacts",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/senders/import": {
      "post": {
        "operationId": "TranslateController_importSenders",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSendersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Count added + updated overrides"
          }
        },
        "summary": "Seed sender overrides from a session's contacts + joined-group members (skips existing)",
        "tags": [
          "translate"
        ]
      }
    },
...
  },
  "info": {
    "title": "FloorLingo API",
    "description": "Self-hosted WhatsApp HTTP API with zh <-> vi shop-floor translation",
    "version": "0.8.17",
    "contact": {
      "name": "FloorLingo",
      "url": "https://github.com/JaplinChen/FloorLingo",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "UpdateTranslateConfigDto": {
        "type": "object",
        "properties": {}
      },
      "LlmProbeDto": {
        "type": "object",
        "properties": {}
      },
      "PreviewTranslateDto": {
        "type": "object",
        "properties": {}
      },
      "GlossaryTermDto": {
        "type": "object",
        "properties": {}
      },
      "ApproveBulkDto": {
        "type": "object",
        "properties": {}
      },
      "SenderEntryDto": {
        "type": "object",
        "properties": {}
      },
      "CategoryDto": {
        "type": "object",
        "properties": {}
      },
      "ImportSendersDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}